Fixing typo in Reflection.Debugging namespace
