+++ Reflection_oM/Debugging/Event.cs
@@ -5,17 +5,21 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BH.oM.Reflection.Debuging
+namespace BH.oM.Reflection.Debugging
 {
-    public class Log : BHoMObject
+    public class Event : BHoMObject
     {
         /***************************************************/
         /**** Properties                                ****/
         /***************************************************/
 
-        public List<Event> AllEvents { get; set; } = new List<Event>();
+        public DateTime Time { get; set; } = DateTime.Now;
 
-        public List<Event> CurrentEvents { get; set; } = new List<Event>();
+        public string StackTrace { get; set; } = "";
+
+        public string Message { get; set; } = "";
+
+        public EventType Type { get; set; } = EventType.Unknown;
 
 
         /***************************************************/
