+++ Reflection_oM/Debugging/EventType.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BH.oM.Reflection.Debuging
+namespace BH.oM.Reflection.Debugging
 {
     public enum EventType
     {
