+++ Reflection_oM/Reflection_oM.csproj
@@ -48,9 +48,9 @@
     <Compile Include="Attributes\NotImplementedAttribute.cs" />
     <Compile Include="Attributes\OutputAttribute.cs" />
     <Compile Include="Attributes\ReleasedAttribute.cs" />
-    <Compile Include="Debuging\Event.cs" />
-    <Compile Include="Debuging\EventType.cs" />
-    <Compile Include="Debuging\Log.cs" />
+    <Compile Include="Debugging\Event.cs" />
+    <Compile Include="Debugging\EventType.cs" />
+    <Compile Include="Debugging\Log.cs" />
     <Compile Include="Interface\IOutput.cs" />
     <Compile Include="Output.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
